[RF] Merge EvaluateFuncs and AnalyticalIntegrals header
For simplicity and code reusability, this PR suggests to merge the
`EvaluateFuncs.h` header and the `AnalyticalIntegrals.h` header into a
single header.
