+++ roofit/roofit/src/RooPoisson.cxx
@@ -18,9 +18,7 @@ Poisson pdf
 #include "RooNaNPacker.h"
 #include "RooBatchCompute.h"
 
-#include "RooFit/Detail/AnalyticalIntegrals.h"
-#include "RooFit/Detail/EvaluateFuncs.h"
-#include "Math/ProbFuncMathCore.h"
+#include <RooFit/Detail/MathFuncs.h>
 
 #include <array>
 
@@ -63,7 +61,7 @@ double RooPoisson::evaluate() const
     np.setPayload(-mean);
     return np._payload;
   }
-  return RooFit::Detail::EvaluateFuncs::poissonEvaluate(k, mean);
+  return RooFit::Detail::MathFuncs::poisson(k, mean);
 }
 
 void RooPoisson::translate(RooFit::Detail::CodeSquashContext &ctx) const
@@ -72,7 +70,7 @@ void RooPoisson::translate(RooFit::Detail::CodeSquashContext &ctx) const
    if (!_noRounding)
       xName = "std::floor(" + xName + ")";
 
-   ctx.addResult(this, ctx.buildCall("RooFit::Detail::EvaluateFuncs::poissonEvaluate", xName, mean));
+   ctx.addResult(this, ctx.buildCall("RooFit::Detail::MathFuncs::poisson", xName, mean));
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -100,7 +98,7 @@ double RooPoisson::analyticalIntegral(Int_t code, const char* rangeName) const
   R__ASSERT(code == 1 || code == 2) ;
 
   RooRealProxy const &integrand = code == 1 ? x : mean;
-  return RooFit::Detail::AnalyticalIntegrals::poissonIntegral(
+  return RooFit::Detail::MathFuncs::poissonIntegral(
      code, mean, _noRounding ? x : std::floor(x), integrand.min(rangeName), integrand.max(rangeName), _protectNegative);
 }
 
@@ -116,7 +114,7 @@ RooPoisson::buildCallToAnalyticIntegral(int code, const char *rangeName, RooFit:
    // Since the integral function is the same for both codes, we need to make sure the indexed observables do not appear
    // in the function if they are not required.
    xName = code == 1 ? "0" : xName;
-   return ctx.buildCall("RooFit::Detail::AnalyticalIntegrals::poissonIntegral", code, mean, xName,
+   return ctx.buildCall("RooFit::Detail::MathFuncs::poissonIntegral", code, mean, xName,
                         integrand.min(rangeName), integrand.max(rangeName), _protectNegative);
 }
