+++ roofit/roofit/src/RooPolynomial.cxx
@@ -34,8 +34,7 @@ RooPolynomial::RooPolynomial(const char*, const char*, RooAbsReal&, const RooArg
 #include "RooMsgService.h"
 #include "RooPolyVar.h"
 
-#include <RooFit/Detail/AnalyticalIntegrals.h>
-#include <RooFit/Detail/EvaluateFuncs.h>
+#include <RooFit/Detail/MathFuncs.h>
 
 #include "TError.h"
 #include <vector>
@@ -109,7 +108,7 @@ double RooPolynomial::evaluate() const
 
    RooPolyVar::fillCoeffValues(_wksp, _coefList);
 
-   return RooFit::Detail::EvaluateFuncs::polynomialEvaluate<true>(_wksp.data(), sz, _lowestOrder, _x);
+   return RooFit::Detail::MathFuncs::polynomial<true>(_wksp.data(), sz, _lowestOrder, _x);
 }
 
 void RooPolynomial::translate(RooFit::Detail::CodeSquashContext &ctx) const
@@ -120,8 +119,7 @@ void RooPolynomial::translate(RooFit::Detail::CodeSquashContext &ctx) const
       return;
    }
 
-   ctx.addResult(
-      this, ctx.buildCall("RooFit::Detail::EvaluateFuncs::polynomialEvaluate<true>", _coefList, sz, _lowestOrder, _x));
+   ctx.addResult(this, ctx.buildCall("RooFit::Detail::MathFuncs::polynomial<true>", _coefList, sz, _lowestOrder, _x));
 }
 
 /// Compute multiple values of Polynomial.
@@ -151,7 +149,7 @@ double RooPolynomial::analyticalIntegral(Int_t code, const char *rangeName) cons
 
    RooPolyVar::fillCoeffValues(_wksp, _coefList);
 
-   return RooFit::Detail::AnalyticalIntegrals::polynomialIntegral<true>(_wksp.data(), sz, _lowestOrder, xmin, xmax);
+   return RooFit::Detail::MathFuncs::polynomialIntegral<true>(_wksp.data(), sz, _lowestOrder, xmin, xmax);
 }
 
 std::string RooPolynomial::buildCallToAnalyticIntegral(Int_t /* code */, const char *rangeName,
@@ -163,6 +161,6 @@ std::string RooPolynomial::buildCallToAnalyticIntegral(Int_t /* code */, const c
    if (!sz)
       return std::to_string(_lowestOrder ? xmax - xmin : 0.0);
 
-   return ctx.buildCall("RooFit::Detail::AnalyticalIntegrals::polynomialIntegral<true>", _coefList, sz, _lowestOrder,
+   return ctx.buildCall("RooFit::Detail::MathFuncs::polynomialIntegral<true>", _coefList, sz, _lowestOrder,
                         xmin, xmax);
 }
