+++ roofit/roofitcore/src/RooEfficiency.cxx
@@ -34,7 +34,7 @@ F may have an arbitrary number of dependents and parameters
 #include "RooStreamParser.h"
 #include "RooArgList.h"
 
-#include <RooFit/Detail/EvaluateFuncs.h>
+#include <RooFit/Detail/MathFuncs.h>
 
 #include "TError.h"
 
@@ -74,7 +74,7 @@ RooEfficiency::RooEfficiency(const RooEfficiency& other, const char* name) :
 double RooEfficiency::evaluate() const
 {
    const int sigCatIndex = _cat->lookupIndex(_sigCatName.Data());
-   return RooFit::Detail::EvaluateFuncs::efficiencyEvaluate(_effFunc, _cat, sigCatIndex);
+   return RooFit::Detail::MathFuncs::efficiency(_effFunc, _cat, sigCatIndex);
 }
 
 int RooEfficiency::getAnalyticalIntegral(RooArgSet &allVars, RooArgSet &analVars, const char * /*rangeName*/) const
@@ -90,7 +90,7 @@ double RooEfficiency::analyticalIntegral(int /*code*/, const char * /*rangeName*
 void RooEfficiency::translate(RooFit::Detail::CodeSquashContext &ctx) const
 {
    int sigCatIndex = _cat->lookupIndex(_sigCatName.Data());
-   ctx.addResult(this, ctx.buildCall("RooFit::Detail::EvaluateFuncs::efficiencyEvaluate", _effFunc, _cat, sigCatIndex));
+   ctx.addResult(this, ctx.buildCall("RooFit::Detail::MathFuncs::efficiency", _effFunc, _cat, sigCatIndex));
 }
 
 std::string RooEfficiency::buildCallToAnalyticIntegral(int /*code*/, const char * /*rangeName*/,
