+++ roofit/roofitcore/src/RooNormalizedPdf.h
@@ -73,8 +73,8 @@ protected:
    void doEval(RooFit::EvalContext &) const override;
    double evaluate() const override
    {
-      // Evaluate() should not be called in the BatchMode, but we still need it
-      // to support printing of the object.
+      // The evaluate() function should not be called in the BatchMode, but we
+      // still need it to support printing of the object.
       return getValV(nullptr);
    }
    double getValV(const RooArgSet * /*normSet*/) const override
