+++ roofit/roofitcore/src/RooRatio.cxx
@@ -27,7 +27,7 @@ Represents the ratio of two RooAbsReal objects.
 #include <RooRealVar.h>
 #include <RooTrace.h>
 
-#include "RooFit/Detail/EvaluateFuncs.h"
+#include "RooFit/Detail/MathFuncs.h"
 
 #include <Riostream.h>
 
@@ -113,7 +113,7 @@ RooRatio::RooRatio(const RooRatio &other, const char *name)
 
 double RooRatio::evaluate() const
 {
-   return RooFit::Detail::EvaluateFuncs::ratioEvaluate(_numerator, _denominator);
+   return RooFit::Detail::MathFuncs::ratio(_numerator, _denominator);
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -126,5 +126,5 @@ void RooRatio::doEval(RooFit::EvalContext &ctx) const
 
 void RooRatio::translate(RooFit::Detail::CodeSquashContext &ctx) const
 {
-   ctx.addResult(this, ctx.buildCall("RooFit::Detail::EvaluateFuncs::ratioEvaluate", _numerator, _denominator));
+   ctx.addResult(this, ctx.buildCall("RooFit::Detail::MathFuncs::ratio", _numerator, _denominator));
 }
