+++ roofit/roofitcore/src/RooRecursiveFraction.cxx
@@ -35,7 +35,7 @@ from a set of recursive fractions: for a given set of input fractions
 #include "RooArgSet.h"
 #include "RooMsgService.h"
 
-#include <RooFit/Detail/EvaluateFuncs.h>
+#include <RooFit/Detail/MathFuncs.h>
 
 ClassImp(RooRecursiveFraction);
 
@@ -90,5 +90,5 @@ double RooRecursiveFraction::evaluate() const
 
 void RooRecursiveFraction::translate(RooFit::Detail::CodeSquashContext &ctx) const
 {
-   ctx.addResult(this, ctx.buildCall("RooFit::Detail::EvaluateFuncs::recursiveFractionEvaluate", _list, _list.size()));
+   ctx.addResult(this, ctx.buildCall("RooFit::Detail::MathFuncs::recursiveFraction", _list, _list.size()));
 }
